backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.18
pydantic==2.0.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.11.1
psycopg2-binary==2.9.6
PyPDF2==3.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0